+++ server/src/main/java/org/apache/iotdb/db/mpp/plan/planner/plan/node/process/last/LastQueryNode.java
@@ -160,7 +160,7 @@ public class LastQueryNode extends MultiChildProcessNode {
     byte needOrderByTimeseries = ReadWriteIOUtils.readByte(byteBuffer);
     Ordering timeseriesOrdering = null;
     if (needOrderByTimeseries == 1) {
-      timeseriesOrdering = Ordering.values()[ReadWriteIOUtils.readByte(byteBuffer)];
+      timeseriesOrdering = Ordering.values()[ReadWriteIOUtils.readInt(byteBuffer)];
     }
     PlanNodeId planNodeId = PlanNodeId.deserialize(byteBuffer);
     return new LastQueryNode(planNodeId, timeFilter, timeseriesOrdering);
